Refactored styled-components into one folder
