+++ client/src/components/employer/EmployerCreditsInfo.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import {
     CreditsContainer,
     TextBox,
-  } from '../styles/navBarInfoStyles';
+  } from '../styles';
 
 class EmployerNavBarInfo extends Component {
 render() {
