+++ client/src/components/employer/EmployerMatches.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-import Progress from '../../containers/Progress';
+
 import { getSeekerMatches, archiveSeeker } from '../../actions';
+import Progress from '../../containers/Progress';
 
 import {
   GridContainer,
