+++ client/src/components/employer/EmployerProfile.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
-import ProfilePicContainer from './ProfilePicContainer';
 import { getUserProfile, updateUserProfile, updateUserPassword } from '../../actions'; 
+
+import ProfilePicContainer from './ProfilePicContainer';
 
 import {
   BodyContainer,
