+++ client/src/components/nav/CreditsInfo.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
+
 import EmployerCreditsInfo from '../employer/EmployerCreditsInfo';
 import SeekerCreditsInfo from '../seeker/SeekerCreditsInfo';
