+++ client/src/components/seeker/SeekerBilling.js
@@ -1,7 +1,9 @@
 import React, { Component } from 'react';
-import BillingForm  from './BillingForm'; 
+
 import {StripeProvider} from 'react-stripe-elements';
 import { Elements } from  'react-stripe-elements';
+
+import BillingForm  from './BillingForm'; 
 
 const stripePublic = process.env.REACT_APP_STRIPE_PUBLIC;
