+++ client/src/components/seeker/SeekerBrowseView.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
+
 import { likeJob } from '../../actions';
 
 import {
