+++ client/src/components/seeker/SeekerMatches.js
@@ -1,8 +1,10 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
-import Progress from '../../containers/Progress';
+import { connect } from 'react-redux';
+
 import { getJobMatches, archiveJob } from '../../actions';
+
+import Progress from '../../containers/Progress';
 
 import {
   GridContainer,
