+++ client/src/components/seeker/SeekerRegister.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
+import { connect } from 'react-redux';
+
 import { registerUser } from '../../actions';
 
 import {
