--- client/src/components/styles/loadingStyles.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const Loading = styled.div`
-    display:flex;
-    justify-content:center;
-    margin-top: 20%;
-
-`;
-
-export default Loading;
