--- client/src/components/styles/navBarInfoStyles.js
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-
-export const CreditsContainer = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  padding-left: 10px;
-  border-bottom: 2px ridge rgba(24, 2, 79);
-`;
-export const TextBox = styled.div`
-  font-size: 24px;
-  padding-right: 25px;
-`;
