+++ client/src/components/styles/old.js
@@ -1,0 +1,97 @@
+// GLOBAL
+import {
+  BodyContainer,
+  ChildContainer,
+  Form,
+  InputContainer,
+  InputTitle,
+  InputBox,
+  InputTextarea,
+  ProfilePic,
+  Notification,
+  ButtonsContainer,
+  ButtonsBox,
+  Button,
+  Link,
+  DropzoneContainer,
+  ModalContainer,
+  Modal,
+  // CARD
+  GridContainer,
+  Card,
+  CardHeader,
+  CardPic,
+  CardName,
+  CardTitle,
+  CardParagraph,
+  CardButton,
+} from './globalStyles';
+// PROFILE
+import {
+  SecurityContainer,
+  ConfirmCheck,
+} from './profileStyles';
+// LANDING
+import {
+  BackgroundContainer,
+  AppTitle,
+  RegisterAndLoginContainer,
+  RegisterContainer,
+  LoginContainer,
+  SeekerOrEmployer,
+  LandingButton,
+} from './landingStyles';
+// BROWSE
+import {
+  NoneLeftMessage,
+  BrowseView,
+  Title,
+  MatchedAndDesired,
+  Paragraph,
+} from './browseStyles';
+
+export {
+  // GLOBAL
+  BodyContainer,
+  ChildContainer,
+  Form,
+  InputContainer,
+  InputTitle,
+  InputBox,
+  InputTextarea,
+  ProfilePic,
+  Notification,
+  ButtonsContainer,
+  ButtonsBox,
+  Button,
+  Link,
+  DropzoneContainer,
+  ModalContainer,
+  Modal,
+  // GLOBAL CARD
+  GridContainer,
+  Card,
+  CardHeader,
+  CardPic,
+  CardName,
+  CardTitle,
+  CardParagraph,
+  CardButton,
+  // PROFILE
+  SecurityContainer,
+  ConfirmCheck,
+  // LANDING
+  BackgroundContainer,
+  AppTitle,
+  RegisterAndLoginContainer,
+  RegisterContainer,
+  LoginContainer,
+  SeekerOrEmployer,
+  LandingButton,
+  // BROWSE
+  NoneLeftMessage,
+  BrowseView,
+  MatchedAndDesired,
+  Title,
+  Paragraph,
+};
