--- client/src/components/styles/profileStyles.js
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-export const ProfilePic = styled.img`
-`;
-export const SecurityContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 70%;
-  margin-left: 20%;
-  padding-top: 5%;
-`;
-export const ConfirmCheck = styled.input`
-  margin-right: 10px;
-  font-size: 16px;
-  box-shadow: 1px 1px 1px rgba(28, 4, 104, .5);
-`;
